db properly configured (i think)
