+++ .gitignore
@@ -1,2 +1,3 @@
 *.json
 __pycache__
+*.db
